refactor: extract CORS config to dedicated file
- Move hardcoded CORS logic from app.ts to config/cors.config.ts
- Make origins configurable via CORS_ORIGINS env var with fallback
